% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickr.R
\name{flickr_context}
\alias{flickr_context}
\title{Return context: full source of this client}
\usage{
flickr_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
