% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickr.R
\name{flickr_geo}
\alias{flickr_geo}
\title{Fetch Flickr geotagged photos feed}
\usage{
flickr_geo(tags = NULL)
}
\arguments{
\item{tags}{Optional comma-separated tags to filter}
}
\value{
tibble: title, link, author, author_id, tags, date_taken, published, image_url
}
\description{
Fetch Flickr geotagged photos feed
}
