% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickr.R
\name{flickr_recent}
\alias{flickr_recent}
\title{Fetch recent public photos from Flickr (no filter)}
\usage{
flickr_recent()
}
\value{
tibble: title, link, author, author_id, tags, date_taken, published, image_url
}
\description{
Fetch recent public photos from Flickr (no filter)
}
