% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickr.R
\name{flickr_search}
\alias{flickr_search}
\title{Search Flickr public photos by tags}
\usage{
flickr_search(tags, tagmode = "any")
}
\arguments{
\item{tags}{Comma-separated tag string (e.g. "nature,landscape")}

\item{tagmode}{Tag matching mode: "any" (OR) or "all" (AND). Default "any".}
}
\value{
tibble: title, link, author, author_id, tags, date_taken, published, image_url
}
\description{
Search Flickr public photos by tags
}
