% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickr.R
\name{flickr_user_photos}
\alias{flickr_user_photos}
\title{Fetch recent public photos from a Flickr user}
\usage{
flickr_user_photos(user_id)
}
\arguments{
\item{user_id}{Flickr user ID (e.g. "12345678@N00")}
}
\value{
tibble: title, link, author, author_id, tags, date_taken, published, image_url
}
\description{
Fetch recent public photos from a Flickr user
}
